+++ apps/webapp/app/entry.server.tsx
@@ -1,8 +1,10 @@
+import * as apihero from "~/services/apihero.server";
 import type { EntryContext } from "@remix-run/node";
 import { RemixServer } from "@remix-run/react";
 import { renderToString } from "react-dom/server";
 import * as Sentry from "@sentry/remix";
 import { prisma } from "./db.server";
+import { env } from "./env.server";
 
 export default function handleRequest(
   request: Request,
@@ -24,8 +26,18 @@ export default function handleRequest(
   });
 }
 
-Sentry.init({
-  dsn: "https://a014169306c748b1adf61875c64b90de:a7fa7bfcc28d43e1bd293e121c677e4a@o4504169280569344.ingest.sentry.io/4504169281880064",
-  tracesSampleRate: 1,
-  integrations: [new Sentry.Integrations.Prisma({ client: prisma })],
-});
+if (process.env.NODE_ENV === "production") {
+  Sentry.init({
+    dsn: env.SENTRY_DSN,
+    tracesSampleRate: 1,
+    integrations: [new Sentry.Integrations.Prisma({ client: prisma })],
+  });
+
+  console.log("🚦 Sentry initialized");
+}
+
+if (process.env.NODE_ENV === "development") {
+  apihero.proxy.start(() => {
+    console.info("🔶 API Hero proxy running");
+  });
+}